@@ -3,6 +3,7 @@ package com.unitbv.datasource;
 import com.unitbv.model.User;
 
 import java.util.*;
+import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -11,7 +12,7 @@ public class UserDataSource {
 
     private List<User> users = new ArrayList<>();
 
-    public UserDataSource(){
+    public UserDataSource() {
         users.add(new User(1, "John", "Wick", 35, "actor"));
         users.add(new User(2, "Jayce", "Lucas", 35, "driver"));
         users.add(new User(3, "Jack", "Spades", 18, "gamer"));
@@ -21,20 +22,20 @@ public class UserDataSource {
         users.add(new User(7, "Mark", "John", 17, "student"));
     }
 
-    public UserDataSource(List<User> users){
+    public UserDataSource(List<User> users) {
         this.users = users;
     }
 
-    public List<User> getAll(){
+    public List<User> getAll() {
         return users;
     }
 
-    public Optional<User> findById(int id){
+    public Optional<User> findById(int id) {
         Optional<User> user = users.stream().filter(u -> u.getId() == id).findFirst();
         return user;
     }
 
-    public String getUsersNamesWithAgeGreaterThanThirty(){
+    public String getUsersNamesWithAgeGreaterThanThirty() {
         String text = users.stream()
                 .filter(u -> u.getAge() > 30)
                 .map(User::getFirstName)
@@ -42,27 +43,27 @@ public class UserDataSource {
         return text;
     }
 
-    public int sumUpUserAgesWhereFirstNameStartsWithJ(){
+    public int sumUpUserAgesWhereFirstNameStartsWithJ() {
         return users.stream()
                 .filter(u -> u.getFirstName().startsWith("J"))
                 .mapToInt(User::getAge)
                 .reduce(0, Integer::sum);
     }
 
-    public User getUserWithHighestAge(){
+    public User getUserWithHighestAge() {
         return users.stream()
                 .sorted((u1, u2) -> Integer.compare(u2.getAge(), u1.getAge()))
                 .findFirst()
                 .orElse(new User());
     }
 
-    public User getUserWithHighestAge_V2(){
+    public User getUserWithHighestAge_V2() {
         return users.stream()
                 .max(Comparator.comparingInt(User::getAge))
                 .orElse(new User());
     }
 
-    public List<User> mergeUserLists(List<User> l1, List<User> l2){
+    public List<User> mergeUserLists(List<User> l1, List<User> l2) {
         return Stream.of(l1, l2)
                 .flatMap(Collection::stream)
                 .collect(Collectors.toList());
@@ -71,78 +72,89 @@ public class UserDataSource {
     // <---------- TO DO ---------->
 
     // Get the full names for all users
-    public List<String> getFullNames(){
+    public List<String> getFullNames() {
         // your code here
-        return new ArrayList<>();
+        return users.stream().map(u -> u.getFirstName() + " " + u.getLastName()).collect(Collectors.toList());
     }
 
     // Get the job of the oldest user
-    public String getJobOfTheOldestUser(){
+    public String getJobOfTheOldestUser() {
         // your code here
-        return "";
+        return users.stream().max(Comparator.comparingInt(User::getAge)).get().getJob();
     }
 
     // Get user (distinct) jobs sorted alphabetically
-    public Set<String> getAllUserJobsSorted(){
+    public Set<String> getAllUserJobsSorted() {
         // your code here
-        return new HashSet<>();
+        return users.stream().map(User::getJob).collect(Collectors.toSet());
     }
 
     // Find user by first name - throw RuntimeException if not found
-    public User findByFirstName(String firstName){
+    public User findByFirstName(String firstName) {
         // your code here
-        return new User();
+        return users.stream().filter(user -> user.getFirstName().equals(firstName)).findFirst().orElseThrow(RuntimeException::new);
     }
 
     // Check if all users are older than the specified age
-    public boolean areAllUsersOlderThan(int age){
+    public boolean areAllUsersOlderThan(int age) {
         // your code here - please try with allMatch/noneMatch
-        return false;
+        return users.stream().allMatch(user -> user.getAge() > age);
     }
 
     // Add a new user - if there is a user with the same id, don't add and throw a RuntimeException
-    public void addUser(User user){
+    public void addUser(User user) {
         // your code here - HINT: use ifPresent() method from Optional
+        users.stream().filter(user1 -> user1.getId() == user.getId()).findFirst().ifPresentOrElse(
+                user1 -> {
+                    throw new RuntimeException();
+                },
+                () -> users.add(user)
+        );
+
     }
 
     // For all students (user.job = "student"), change the job to "graduate" and add 5 years to their age
-    public void changeAllStudentsJobsAndAges(){
+    public void changeAllStudentsJobsAndAges() {
         // your code here
+        users.stream().filter(user -> user.getJob().equals("student")).forEach(user -> {
+            user.setAge(user.getAge() + 5);
+            user.setJob("graduate");
+        });
     }
 
     // Count users that have the given Job
-    public long countUsersHavingTheSpecifiedJob(String job){
+    public long countUsersHavingTheSpecifiedJob(String job) {
         // your code here
-        return 0;
+        return users.stream().filter(user -> user.getJob().equals(job)).count();
     }
 
     // Get a map where the key is the user id and the value is the User object itself
-    public Map<Integer, User> getMapOfUsers(){
+    public Map<Integer, User> getMapOfUsers() {
         // your code here
-        return new HashMap<>();
+        return users.stream().collect(Collectors.toMap(User::getId, Function.identity()));
     }
 
     // Get a predicate for filtering by the given name - applies to both firstName and lastName
-    public Predicate<User> getPredicateForFilteringByName(String name){
+    public Predicate<User> getPredicateForFilteringByName(String name) {
         // your code here
-        return null;
+        return p -> p.getFirstName().equals(name) || p.getLastName().equals(name);
     }
 
     // Get a comparator for User type - compare by age ascending, then by job alphabetically
-    public Comparator<User> getUserComparator(){
+    public Comparator<User> getUserComparator() {
         // your code here
-        return null;
+        return Comparator.comparingInt(User::getAge).thenComparing(User::getJob);
     }
 
     // Filter users using the given Predicate
-    public List<User> filterUsers(Predicate<? super User> predicate){
+    public List<User> filterUsers(Predicate<? super User> predicate) {
         return users.stream()
                 .filter(predicate)
                 .collect(Collectors.toList());
     }
 
     // Sort users using the given Comparator
-    public List<User> sortUsers(Comparator<? super User> comparator){
+    public List<User> sortUsers(Comparator<? super User> comparator) {
         return users.stream()
                 .sorted(comparator)
                 .collect(Collectors.toList());
