@@ -1,5 +1,6 @@
 package com.unitbv.events.service;
 
+import com.unitbv.events.event.CarForRentEvent;
 import com.unitbv.events.model.Car;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -15,6 +16,7 @@ public class PersonService {
 
     public void rentCar(String model) {
         final Car car = new Car(model);
-        //create a new event and publish car on event
+        CarForRentEvent carEvent = new CarForRentEvent(car);
+        eventPublisher.publishEvent(carEvent);
     }
 }
