@@ -1,10 +1,20 @@
 package com.unitbv.profiles.stereotype;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.annotation.*;
 
 @Configuration
-@ComponentScan
+@ComponentScan("com.unitbv.profiles.stereotype")
+@PropertySource("classpath:profiles.properties")
 public class AppConfig {
+    @Bean
+    @Profile({"dev", "default"})
+    DevDataSourceConfig devDataSourceConfig() {
+        return new DevDataSourceConfig();
+    }
+
+    @Bean
+    @Profile("prod")
+    ProdDataSourceConfig prodDataSourceConfig() {
+        return new ProdDataSourceConfig();
+    }
 }
